# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zerizeha API
  version: "1.0"
  description: Zerizeha backend API
servers:
  - url: /
paths:
...
  /api/voice/channels/{id}/join:
    post:
      operationId: JoinVoiceChannel
      summary: join voice channel (presence)
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/voice/channels/{id}/members:
    get:
      operationId: ListVoiceMembers
      summary: list voice members (presence)
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoiceMember'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/channels/voice/{id}/webrtc/bootstrap:
    post:
      operationId: VoiceWebRTCBootstrap
      summary: bootstrap janus webrtc signaling for voice channel
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebRTCBootstrapResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/voice/leave:
    post:
      operationId: LeaveVoice
      summary: leave active voice channel (presence)
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/voice/heartbeat:
    post:
      operationId: VoiceHeartbeat
      summary: voice presence heartbeat
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
  /api/channels/{id}:
    get:
      operationId: GetChannelByID
      summary: get channel by id
      tags:
        - Channels
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
    patch:
      operationId: UpdateChannel
      summary: update channel
      tags:
        - Channels
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelToUpdate'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
    delete:
      operationId: DeleteChannel
      summary: delete channel
      tags:
        - Channels
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    ErrorMap:
      type: object
      additionalProperties:
        type: string
...
    Channel:
      type: object
      required:
        - id
        - author_id
        - space_id
        - name
        - channel_type
        - created_at
      properties:
        id:
          type: string
        author_id:
          type: string
        space_id:
          type: string
        name:
          type: string
        channel_type:
          type: string
          enum: [text, voice]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...
    ChannelToUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
      additionalProperties: false
...
    VoiceMember:
      type: object
      required: [id, username, is_admin]
      properties:
        id:
          type: string
        username:
          type: string
        is_admin:
          type: boolean
        avatar_url:
          type: string
          nullable: true
      additionalProperties: false
    WebRTCPublisher:
      type: object
      required: [feed_id]
      properties:
        feed_id:
          type: string
        display:
          type: string
          nullable: true
      additionalProperties: false
    WebRTCBootstrapResponse:
      type: object
      required: [connection_id, room_id, self_feed_id, publishers]
      properties:
        connection_id:
          type: string
        room_id:
          type: string
        self_feed_id:
          type: string
        publishers:
          type: array
          items:
            $ref: '#/components/schemas/WebRTCPublisher'
      additionalProperties: false